requirements.txt
asgiref==3.5.2
Django==4.0.5
sqlparse==0.4.2
tzdata==2022.1
mysqlclient==2.1.1
djangorestframework
gunicorn==20.1.0
django-cors-headers
requests==2.28.0
google-cloud-storage
djangorestframework-simplejwt
python-dotenv==0.20.0
django-allauth==0.51.0
whitenoise
django-allauth==0.51.0
psycopg2-binary==2.9.6
psycopg2==2.9.6